Move client-like functions into new client.py
