+++ db_rest_api/api.py
@@ -4,7 +4,7 @@ import logging
 from flask import Flask, request, abort, jsonify, Response
 from flask_compress import Compress
 
-from indra.db.util import get_statements_by_gene_role_type, \
+from indra.db.client import get_statements_by_gene_role_type, \
     get_statements_by_paper
 from indra.statements import make_statement_camel
 from indra.databases import hgnc_client
