+++ indra/db/util.py
@@ -2,18 +2,13 @@ from __future__ import absolute_import, print_function, unicode_literals
 from builtins import dict, str
 
 __all__ = ['get_defaults', 'get_primary_db', 'insert_agents', 'insert_pa_stmts',
-           'insert_db_stmts', 'get_abstracts_by_pmids', 'get_auth_xml_pmcids',
-           'get_statements_by_gene_role_type', 'get_statements',
-           'make_stmts_from_db_list']
+           'insert_db_stmts', 'get_auth_xml_pmcids', 'make_stmts_from_db_list']
 
 import re
 import json
-import zlib
 import logging
 from sqlalchemy import func
-from indra.databases import hgnc_client
 from indra.util.get_version import get_version
-from indra.util import unzip_string
 from indra.statements import Complex, SelfModification, ActiveForm,\
     stmts_from_json, Conversion, Translocation, Evidence, Statement
 from .database_manager import DatabaseManager, IndraDatabaseError, texttypes
@@ -279,41 +274,6 @@ def insert_pa_stmts(db, stmts, verbose=False):
     return
 
 
-def get_abstracts_by_pmids(db, pmid_list, unzip=True):
-    """Return abstracts given a list of PMIDs from the database
-
-    Parameters
-    ----------
-    db : :py:class:`DatabaseManager`
-        Reference to the DB to query
-    pmid_list : list[str]
-        A list of PMIDs whose abstracts are to be returned
-    unzip : Optional[bool]
-        If True, the compressed output is decompressed into clear text.
-        Default: True
-
-    Returns
-    -------
-    abstracts : dict
-        A dictionary whose keys are PMIDs with each value being the
-        the corresponding abstract
-    """
-    abst_list = db.filter_query(
-        [db.TextRef, db.TextContent],
-        db.TextContent.text_ref_id == db.TextRef.id,
-        db.TextContent.text_type == 'abstract',
-        db.TextRef.pmid.in_(pmid_list)
-        ).all()
-    if unzip:
-        def unzip_func(s):
-            return zlib.decompress(s, zlib.MAX_WBITS + 16).decode('utf-8')
-    else:
-        def unzip_func(s):
-            return s
-    abstracts = {r.pmid: unzip_func(c.content) for (r, c) in abst_list}
-    return abstracts
-
-
 def get_auth_xml_pmcids(db):
     tref_list = db.filter_query(
         [db.TextRef, db.TextContent],
@@ -322,198 +282,6 @@ def get_auth_xml_pmcids(db):
         db.TextContent.source == 'pmc_auth'
         )
     return [tref.pmcid for tref in tref_list]
-
-
-#==============================================================================
-# Below are some functions that are useful for getting raw statements from the
-# database at various levels of abstraction.
-#==============================================================================
-
-def get_statements_by_gene_role_type(agent_id=None, agent_ns='HGNC-SYMBOL',
-                                     role=None, stmt_type=None, count=1000,
-                                     db=None, do_stmt_count=True,
-                                     preassembled=True):
-    """Get statements from the DB by stmt type, agent, and/or agent role.
-
-    Parameters
-    ----------
-    agent_id : str
-        String representing the identifier of the agent from the given
-        namespace. Note: if the agent namespace argument, `agent_ns`, is set
-        to 'HGNC-SYMBOL', this function will treat `agent_id` as an HGNC gene
-        symbol and perform an internal lookup of the corresponding HGNC ID.
-        Default is 'HGNC-SYMBOL'.
-    agent_ns : str
-        Namespace for the identifier given in `agent_id`.
-    role : str
-        String corresponding to the role of the agent in the statement.
-        Options are 'SUBJECT', 'OBJECT', or 'OTHER' (in the case of `Complex`,
-        `SelfModification`, and `ActiveForm` Statements).
-    stmt_type : str
-        Name of the Statement class.
-    count : int
-        Number of statements to retrieve in each batch (passed to
-        :py:func:`get_statements`).
-    db : :py:class:`DatabaseManager`
-        Optionally specify a database manager that attaches to something
-        besides the primary database, for example a local databse instance.
-    do_stmt_count : bool
-        Whether or not to perform an initial statement counting step to give
-        more meaningful progress messages.
-    preassembled : bool
-        If true, statements will be selected from the table of pre-assembled
-        statements. Otherwise, they will be selected from the raw statements.
-        Default is True.
-
-    Returns
-    -------
-    list of Statements from the database corresponding to the query.
-    """
-    if db is None:
-        db = get_primary_db()
-
-    if preassembled:
-        Statements = db.PAStatements
-        Agents = db.PAAgents
-    else:
-        Statements = db.Statements
-        Agents = db.Agents
-
-    if not (agent_id or role or stmt_type):
-        raise ValueError('At least one of agent_id, role, or stmt_type '
-                         'must be specified.')
-    clauses = []
-    if agent_id and agent_ns == 'HGNC-SYMBOL':
-        hgnc_id = hgnc_client.get_hgnc_id(agent_id)
-        if not hgnc_id:
-            logger.warning('Invalid gene name: %s' % agent_id)
-            return []
-        clauses.extend([Agents.db_name.like('HGNC'),
-                        Agents.db_id.like(hgnc_id)])
-    elif agent_id:
-        clauses.extend([Agents.db_name.like(agent_ns),
-                        Agents.db_id.like(agent_id)])
-    if role:
-        clauses.append(Agents.role == role)
-    if agent_id or role:
-        clauses.append(Agents.stmt_id == Statements.id)
-    if stmt_type:
-        clauses.append(Statements.type == stmt_type)
-    stmts = get_statements(clauses, count=count, do_stmt_count=do_stmt_count,
-                           db=db, preassembled=preassembled)
-    return stmts
-
-
-def get_statements_by_paper(id_val, id_type='pmid', count=1000, db=None,
-                            do_stmt_count=True):
-    """Get the statements from a particular paper.
-
-    Note: currently this can only retrieve raw statements, because of the
-    partially implemented configuration of the pre-assembled Statement table.
-
-    Parameters
-    ----------
-    id_val : int or str
-        The value of the id for the paper whose statements you wish to retrieve.
-    id_type : str
-        The type of id used (default is pmid). Options include pmid, pmcid, doi,
-        pii, url, or manuscript_id. Note that pmid is generally the best means
-        of getting a paper.
-    count : int
-        Number of statements to retrieve in each batch (passed to
-        :py:func:`get_statements`).
-    db : :py:class:`DatabaseManager`
-        Optionally specify a database manager that attaches to something
-        besides the primary database, for example a local databse instance.
-    do_stmt_count : bool
-        Whether or not to perform an initial statement counting step to give
-        more meaningful progress messages.
-
-    Returns
-    -------
-    A list of Statements from the database corresponding to the paper id given.
-    """
-    if db is None:
-        db = get_primary_db()
-
-    # Get the text ref id(s)
-    if id_type in ['trid']:
-        trid_list = [int(id_val)]
-    else:
-        id_types = ['pmid', 'pmcid', 'doi', 'pii', 'url', 'manuscript_id']
-        if id_type not in id_types:
-            raise ValueError('id_type must be one of: %s' % str(id_types))
-        constraint = (getattr(db.TextRef, id_type) == id_val)
-        trid_list = [trid for trid, in db.select_all(db.TextRef.id, constraint)]
-
-    if not trid_list:
-        return None
-
-    stmts = []
-    for trid in trid_list:
-        clauses = [
-            db.TextContent.text_ref_id == trid,
-            db.Readings.text_content_id == db.TextContent.id,
-            db.Statements.reader_ref == db.Readings.id
-        ]
-        stmts.extend(get_statements(clauses, count=count, preassembled=False,
-                                    do_stmt_count=do_stmt_count, db=db))
-    return stmts
-
-
-def get_statements(clauses, count=1000, do_stmt_count=True, db=None,
-                   preassembled=True):
-    """Select statements according to a given set of clauses.
-
-    Parameters
-    ----------
-    clauses : list
-        list of sqlalchemy WHERE clauses to pass to the filter query.
-    count : int
-        Number of statements to retrieve and process in each batch.
-    do_stmt_count : bool
-        Whether or not to perform an initial statement counting step to give
-        more meaningful progress messages.
-    db : :py:class:`DatabaseManager`
-        Optionally specify a database manager that attaches to something
-        besides the primary database, for example a local database instance.
-    preassembled : bool
-        If true, statements will be selected from the table of pre-assembled
-        statements. Otherwise, they will be selected from the raw statements.
-        Default is True.
-
-    Returns
-    -------
-    list of Statements from the database corresponding to the query.
-    """
-    if db is None:
-        db = get_primary_db()
-
-    stmts_tblname = 'pa_statements' if preassembled else 'statements'
-
-    stmts = []
-    q = db.filter_query(stmts_tblname, *clauses)
-    if do_stmt_count:
-        logger.info("Counting statements...")
-        num_stmts = q.count()
-        logger.info("Total of %d statements" % num_stmts)
-    db_stmts = q.yield_per(count)
-    subset = []
-    total_counter = 0
-    for stmt in db_stmts:
-        subset.append(stmt)
-        if len(subset) == count:
-            stmts.extend(make_stmts_from_db_list(subset))
-            subset = []
-        total_counter += 1
-        if total_counter % count == 0:
-            if do_stmt_count:
-                logger.info("%d of %d statements" % (total_counter, num_stmts))
-            else:
-                logger.info("%d statements" % total_counter)
-
-    stmts.extend(make_stmts_from_db_list(subset))
-    return stmts
 
 
 def make_stmts_from_db_list(db_stmt_objs):
